COMP: add help to find SlicerOpenCV on factory machines
On the factory dashboards, this example extension wasn't building because
it wasn't finding the SlicerOpenCV extension that it depended on. As a quick
fix, this commit provides a find file that will look in the most likely place
on the dashboard for the SlicerOpenCV build directory and include the config
file to define needed variables for linking against OpenCV.
For developers working locally, it looks for the build directory one or two levels above
and in an OpenCV subdirectory. If not found, the developer can still manually set
SlicerOpenCV_DIR at configure time.

https://github.com/SBU-BMI/SlicerOpenCV/issues/6